rename : convert -> util (#30)
